new files

+++ personal.profile
@@ -9,19 +9,29 @@
 
 # input mode
 set -o vi
+umask 0007
 
 # terminal line settings
-stty erase '^?'
-stty intr '^C' 
-stty kill '^U' 
-stty stop '^S'
-stty susp '^Z'
-stty werase '^W'
-
-# command line _eye candy_
-CCLEAR="\033[00m"; CWHITE="\033[01;37m"
-CRED="\033[01;31m"; CYELLOW="\033[01;33m"
-CBLUE="\033[01;34m"; CGRAY="\033[01;30m"
+stty 2> /dev/null > /dev/null 
+if [ "$?" = "0" ]
+then
+   stty erase '^?'
+   stty intr '^C' 
+   stty kill '^U' 
+   stty stop '^S'
+   stty susp '^Z'
+   stty werase '^W'
+   
+   # command line _eye candy_
+   CCLEAR="\033[00m"; CWHITE="\033[01;37m"
+   CRED="\033[01;31m"; CYELLOW="\033[01;33m"
+   CBLUE="\033[01;34m"; CGRAY="\033[01;30m"
+else
+   # command line _eye candy_
+   CCLEAR=""; CWHITE=""
+   CRED=""; CYELLOW=""
+   CBLUE=""; CGRAY=""
+fi
 
 # functions
 java15 () {
@@ -40,18 +50,22 @@ java12 () {
    PATH=/opt/java1.2/bin:${LOCALPATH}
 }
 
+apache20 () {
+   [ -d /opt/hpws/apache/bin ] && LOCALPATH=/opt/hpws/apache/bin:${LOCALPATH}
+}
+
 newstyle () {
    # command line _eye candy_
    CCLEAR="\033[00m"; CWHITE="\033[01;37m"
    CRED="\033[01;31m"; CYELLOW="\033[01;33m"
    CBLUE="\033[01;34m"; CGRAY="\033[01;30m"
 
-   export PS1="$(echo "${CBLUE}${USER}${CWHITE}@${IPLN}(${CRED}${HOST}${CWHITE})${CCLEAR}") \${PWD##*/} $ "
-   export PS2="  -> "
+   export PS1="$(echo "${CBLUE}${USER}${CCLEAR}@[${CGRAY}${MYIP}(${HOST}${CCLEAR})]") \${PWD##*/} $> "
+   export PS2=" > "
 }
 
 oldstyle () {
-   export PS1="$(echo "\n${CGRAY}${IPLN}(${HOST})${CCLEAR}:${CWHITE}\${PWD}\n${CBLUE}${USER}${CCLEAR}$> ")"
+   export PS1="$(echo "${CCLEAR}\n[${CGRAY}${MYIP}(${HOST})${CCLEAR}]:${CWHITE}\${PWD}\n${CBLUE}${USER}${CCLEAR} $> ")"
    export PS2=" > "
 }
 
@@ -81,23 +95,27 @@ alias lr='lt -r'
 alias pw='pwd'
 alias domains='cd ~/bea/user_projects/domains'
 
-# Commands
-PING=`whereis ping | awk '{print $2}'`
-
 # terminal type
 export EDITOR=vi
 export TERM="xterm"
 # history file
-export HISTSIZE=300
+export HISTSIZE=500
 export HISTCONTROL=ignoredups
 # otros
 export PATH=${LOCALPATH}
 export HOST=`hostname | tr "[:upper:]" "[:lower:]" | sed -e "s/m.*hp//g"`
-export IPLN=`${PING} $(hostname) -c1 | awk '/bytes from/{gsub(":","",$4);print $4}' > /dev/null 2>&1`
 export MANPATH=$HOME/monopse:$MANPATH
-# 
-[ "$SSH_CONNECTION" != "" ] && IPLN=`echo $SSH_CONNECTION | cut -f3 -d" "`
+
+unset USERNAME
+
+# get IP Address
+MYIP=`/usr/sbin/ping $(hostname) -c1 2> /dev/null | awk '/bytes from/{gsub(":","",$4);print $4}' `
+[ "x$MYIP" = "x" ] && MYIP=`echo $SSH_CONNECTION 2> /dev/null | awk '{print $3}' | sed -e "s/.*://g;s/ .*//g"`
+[ "x$MYIP" = "x" ] && MYIP=`/usr/sbin/ifconfig lan0 2> /dev/null | grep "inet" | sed -e "s/.*inet //g;s/netmask.*//g"`
+[ "x$MYIP" = "x" ] && MYIP=`/usr/sbin/ifconfig lan1 2> /dev/null | grep "inet" | sed -e "s/.*inet //g;s/netmask.*//g"`
 
 # Cursor and profile
+apache20
+java14
 oldstyle
